requirements.txt
luma.core
Pillow>=10.0
requests
websocket-client
pyzmq
